pyproject.toml
[tool.poetry]
name = "pzero-improves-zar"
version = "0.1.0"
description = ""
authors = ["ryuto konno <pxx.rai@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "1.19.1"
torch = "1.5.1"
logzero = "1.5.0"
transformers = "3.4.0"
torchtext = "0.6.0"
ipadic = "^1.0.0"
fugashi = "^1.1.1"
mojimoji = "0.0.11"
termcolor = "^1.1.0"
PyYAML = "^6.0"
pytorch-warmup = "^0.0.4"
tensorboardX = "^2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
